pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["HoldCarter <holdcarter@holdcarter.ru>"]
readme = "README.md"
packages = [{include = "gendiff"}]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff_run:main"

[tool.poetry.dependencies]
python = "^3.10"
pytest-cov = "^4.0.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.3.0"
coverage = "^7.2.3"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
